fix: naming scheme change and type change
